requirements.txt
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2==2.9.10
python-dotenv==1.1.0
reportlab==4.0.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.5.0
Werkzeug==3.1.3